fix: rename filename to achieve repository standards
